requirements.txt
mysql-connector
boto3
botocore.exceptions
botocore.retries
flask
flask_jwt
flask_cors